// src/app/enums/component-path-enum.enum.ts
export enum ComponentPathEnum {
    Login = 'login',
    Login2= 'Login2',
    Login3= 'Login3',

    SideBar = 'sidebar',
    DashBoard = 'dashboard',
    RealTimeView = 'RealTimeView',
    SingleBmsDataView = 'SingleBmsDataView',
    UnderDevelopment = 'underdevelopment',
    ReportHistorical = 'reportHistorical',
    ReportDayWise = 'reportDayWise',
    ReportMonthly = 'reportMonthly',

    
    SiteLocation = 'SiteLocation',
    ManufacturerDetails='Manufacturer',
    AlarmsReport='Alarms',

    RawData= 'rw',
    
    Testing1 = 't1',
    Testing2 = 't2',
    Testing3 = 't3',
    Testing4 = 'testing4',
    Testing5 = 'testing5'

}